Add "tests/test_ntoa_assert"
